move code from app class to FileInfo
